[issue] The name of the file can't be the timestamp, as there is a possibility that a post is submitted with the same timestamp.
[solution] TO solve this, and to introduce consistency, the post is stored with the filename
as a hash, of the userid and the present time stamp.

As the same user can never submit two blog posts in the same microsecond, and even if he/she
does, he is simply spamming or testing the system, the hash can never be the same. (A slight difference in the string being encoded completly changes the hash.)