Heavily redesigned the header file.
Now, depending on whether the user is logged in or not, the login or the logour button is showed. 

Further if the user is logged out then instead (Username is logged in) nothing is shown.

Also, fixed an issue due to which any user who was not logged in and came to blogcont was infinitely redirected to the same page again and again.